+++ introspection/remote.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 
 	"github.com/Yamashou/gqlgenc/v3/client"
+	"github.com/Yamashou/gqlgenc/v3/internal/transport"
 
 	"github.com/vektah/gqlparser/v2/ast"
 	"github.com/vektah/gqlparser/v2/validator"
@@ -18,7 +19,7 @@ import (
 func LoadRemoteSchema(ctx context.Context, endpoint string, header http.Header) (*ast.Schema, error) {
 	var options []client.Option
 	if len(header) > 0 {
-		options = append(options, client.WithRoundTripper(client.NewHeaderTransport(func(context.Context) http.Header {
+		options = append(options, client.WithRoundTripper(transport.NewHeader(func(context.Context) http.Header {
 			return header
 		})))
 	}